reconstructure

+++ network/henet.py
@@ -13,10 +13,10 @@ class HENet(nn.Module):
         net = Backbone(backbone)
         n_features = net.bb_info['in_features']
         self.backbone = nn.Sequential(*list(net.backbone.children())[:-2])
-        self.reduce_channel = nn.Conv2d(n_features, n_class * block_expansion, 1, 1)
+        self.reduce_channel = nn.Conv2d(n_features, int(n_class * block_expansion), 1, 1)
         self.HE_block = HEMaxBlock(beta)
         self.pool = nn.AdaptiveAvgPool2d((1, 1))
-        self.fc = nn.Linear(n_class * block_expansion, n_class)
+        self.fc = nn.Linear(int(n_class * block_expansion), n_class)
 
         self.__set_params()
